fix : Altibase 에 맞게 문법 수정

+++ admin/src/main/java/com/example/admin/payment_history/dto/PaymentHistory.java
@@ -11,7 +11,7 @@ public class PaymentHistory { // TB_AUTH_INFO 에서 가져올 데이터
     private String ctn;
     private String transactionDt;
     private String cancelNotificationDt;
-    private Integer total; // itemPrice + tax
+    private Double total; // itemPrice + tax
     private String itemName;
     private String merchantName;
 }
